requirements.txt
mysql-connector-python==9.5.0
pytest==8.3.4
python-dotenv==1.0.1